+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Added new validator to PensionValidator to check for amount lower than 10 millions.
 
 ## [2.13.0] - 2024-07-30Z
 - Updated internal calculation to force Scottish Rate if consumer set `userPaysScottishTax` to true.